code/train_rag/requirements.txt
transformers==4.51.3
datasets
huggingface_hub
torch
faiss-gpu-cu12
numpy==1.26.4
tqdm
sentence-transformers
accelerate>=0.26.0